requirements.txt
requests
pytz
qrcode[pil]
pynacl
pycryptodome

